requirements.txt

python-dotenv==1.0.1
langchain-google-genai==2.0.8
pymssql==2.2.11
pyodbc==5.0.1
python-multipart==0.0.9
langchain-core==0.3.65
langchain-community>=0.0.19
langchain==0.3.25
google-generativeai==0.8.5
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.11.7
sqlalchemy==2.0.28

